Sistema de pulo e movimento concluidos
