+++ Assets/GameFolder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 55c3be7fd6d27ca45b94291c505515b4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
