+++ Assets/GameFolder/Player/Material.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42db3e63be1a0da4cb82a507fcf6d3d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
