+++ Assets/GameFolder/Player/Scripts/Dash.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c117cbc89de2d2840903a5dd426b1c64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
