+++ Assets/GameFolder/Player/Scripts/GroundCheck.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GroundCheck : MonoBehaviour
+{
+    public int nJumping;
+
+
+    void OnTriggerEnter2D(Collider2D other)
+    {
+        if (other.gameObject.layer == 6)
+        {
+            nJumping = 2;
+        }
+    }
+}
