+++ Assets/GameFolder/Player/Scripts/GroundCheck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47e13a2536fcfc040b7bd999e0ddb09e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
