+++ Assets/GameFolder/Player/Scripts/WallJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6db1d6319131554bb6955b73da2eb77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
